Many changes

+++ cloud/custom/oraclelinux/install-common.sh
@@ -11,8 +11,9 @@ fi
 yum install -y -q zip unzip bzip2 xz tig jq sshpass screen python2-httpie moreutils
 
 yum install -y -q yum-utils
-yum-config-manager --add-repo http://download.opensuse.org/repositories/shells:/fish:/release:/3/RHEL_7/shells:fish:release:3.repo
-yum install -y -q fish
+if ! yum repolist all | grep -s -q "^shells_fish_release_3"; then
+    yum install -y -q fish
+fi
 
 if [[ -n ${CUSTOM_UPDATE} && "${CUSTOM_UPDATE}" == "true" ]]; then
     yum -y -q update
